renamed fetchUtil to FetchWrapper
